Add files via upload
Implement a program that runs a plurality election, perhaps the simplest way to hold an election (also known as “first-past-the-post” or “winner take all”). 
In the plurality vote, every voter gets to vote for one candidate. At the end of the election, whichever candidate has the greatest number of votes is declared the winner of the election.