Add files via upload
Not the best example of a good code, but it's good enough 'cause I gave up.

A program that runs a runoff election; a ranked-choice system, where voters can vote for more than one candidate. 
Instead of just voting for their top choice, they can rank the candidates in order of preference. 
If any candidate has a majority (more than 50%) of the first preference votes, that candidate is declared the winner of the election.
If no candidate has more than 50% of the vote, then an “instant runoff” occurs. The candidate who received the fewest number of votes is eliminated from the election, and anyone who originally chose that candidate as their first preference now has their second preference considered. Why does it this way? Effectively, this simulates what would have happened if the least popular candidate had not been in the election, to begin with.
The process repeats: if no candidate has a majority of the votes, the last-place candidate is eliminated, and anyone who voted for them will instead vote for their next preference (who hasn’t themselves already been eliminated). Once a candidate has a majority, that candidate is declared the winner.